+++ src/editor/gui/elements/StyledTreeItem.tsx
